🚚[Rename] list/ui 아래로 위치 변경
