+++ src/components/features/record/list/ui/RecordFilterBar.tsx
@@ -1,11 +1,9 @@
 import { Clock, Flame, Globe, Users } from "lucide-react";
 import ToggleGroup, { ToggleOption } from "@/components/ui/button/ToggleGroup";
 import CycleButton from "@/components/ui/button/CycleButton";
+import { FeedMode, SortMode } from "../types";
 
-export type SortMode = "recent" | "best";
-export type FeedMode = "all" | "following";
-
-type Props = {
+type RecordFilterBarProps = {
   sortMode: SortMode;
   feedMode: FeedMode;
   onSortChange: (mode: SortMode) => void;
@@ -14,12 +12,20 @@ type Props = {
 
 const FEED_OPTIONS: ToggleOption<FeedMode>[] = [
   { value: "all", label: "전체", icon: <Globe className="w-3.5 h-3.5" /> },
-  { value: "following", label: "팔로잉", icon: <Users className="w-3.5 h-3.5" /> },
+  {
+    value: "following",
+    label: "팔로잉",
+    icon: <Users className="w-3.5 h-3.5" />,
+  },
 ];
 
 const SORT_OPTIONS: ToggleOption<SortMode>[] = [
-  { value: "recent", label: "최신순", icon: <Clock className="w-3.5 h-3.5" /> },
-  { value: "best", label: "인기순", icon: <Flame className="w-3.5 h-3.5" /> },
+  { value: "latest", label: "최신순", icon: <Clock className="w-3.5 h-3.5" /> },
+  {
+    value: "popular",
+    label: "인기순",
+    icon: <Flame className="w-3.5 h-3.5" />,
+  },
 ];
 
 export default function RecordFilterBar({
@@ -27,7 +33,7 @@ export default function RecordFilterBar({
   feedMode,
   onSortChange,
   onFeedChange,
-}: Props) {
+}: RecordFilterBarProps) {
   return (
     <div className="flex items-center justify-between">
       <ToggleGroup
@@ -35,7 +41,11 @@ export default function RecordFilterBar({
         value={feedMode}
         onChange={onFeedChange}
       />
-      <CycleButton options={SORT_OPTIONS} value={sortMode} onChange={onSortChange} />
+      <CycleButton
+        options={SORT_OPTIONS}
+        value={sortMode}
+        onChange={onSortChange}
+      />
     </div>
   );
 }
